% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactlyone.R
\name{exactlyone}
\alias{exactlyone}
\title{exactlyone}
\usage{
exactlyone(p)
}
\arguments{
\item{p}{a vector contains probabilities pi}
}
\value{
the probability of it
}
\description{
Compute the probability of exactly one of these events occurring using R
}
\examples{
\dontrun{
p <- c(0.1,0.2,0.3,0.4)
exactlyone(p=p)
}
}
