% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findols.R
\name{findols}
\alias{findols}
\title{findols}
\usage{
findols(x)
}
\arguments{
\item{x}{a matrix, each row of which is observations of corresponding objects}
}
\value{
a vector contains the indices of the most deviant observation in each row
}
\description{
identify the most deviant observation for each row of x using R
}
\examples{
\dontrun{
rs <- matrix(round(rnorm(60)*10,digits = 2), nrow = 4, byrow = TRUE)
findols(rs)
}
}
